// types/HeadingTypeStyle.ts
export enum HeadingTypeStyle {
    xl = 'heading-xl',
    lg = 'heading-lg',
    md = 'heading-md',
    mdSemibold = 'heading-md-semibold',
    sm = 'heading-sm',
    smSemibold = 'heading-sm-semibold',
    xs = 'heading-xs',
    xsSemibold = 'heading-xs-semibold',
    bodyLg = 'body-lg',
    bodyMd = 'body-md',
    bodySm = 'body-sm',
  }